+++ api/argoApplication/ArgoApplicationRestHandler.go
@@ -21,8 +21,8 @@ import (
 	"errors"
 	"github.com/devtron-labs/devtron/api/restHandler/common"
 	"github.com/devtron-labs/devtron/pkg/auth/authorisation/casbin"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read"
 	"go.uber.org/zap"
 	"net/http"
 	"strconv"
