+++ api/helm-app/HelmAppRestHandler.go
@@ -25,11 +25,11 @@ import (
 	"github.com/devtron-labs/devtron/pkg/appStore/installedApp/service"
 	"github.com/devtron-labs/devtron/pkg/appStore/installedApp/service/EAMode"
 	clusterBean "github.com/devtron-labs/devtron/pkg/cluster/bean"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/helper"
 	clientErrors "github.com/devtron-labs/devtron/pkg/errors"
 	"github.com/devtron-labs/devtron/pkg/fluxApplication"
 	bean2 "github.com/devtron-labs/devtron/pkg/k8s/application/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/helper"
 	"github.com/devtron-labs/devtron/pkg/pipeline"
 	"net/http"
 	"strconv"
