+++ api/k8s/application/k8sApplicationRestHandler.go
@@ -36,14 +36,14 @@ import (
 	"github.com/devtron-labs/devtron/pkg/auth/authorisation/casbin"
 	"github.com/devtron-labs/devtron/pkg/auth/user"
 	bean4 "github.com/devtron-labs/devtron/pkg/cluster/environment/bean"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/helper"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read"
 	clientErrors "github.com/devtron-labs/devtron/pkg/errors"
 	"github.com/devtron-labs/devtron/pkg/fluxApplication"
 	"github.com/devtron-labs/devtron/pkg/k8s"
 	application2 "github.com/devtron-labs/devtron/pkg/k8s/application"
 	bean2 "github.com/devtron-labs/devtron/pkg/k8s/application/bean"
 	bean3 "github.com/devtron-labs/devtron/pkg/k8s/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/helper"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read"
 	"github.com/devtron-labs/devtron/pkg/terminal"
 	"github.com/devtron-labs/devtron/util"
 	"github.com/devtron-labs/devtron/util/rbac"
