+++ cmd/external-app/wire_gen.go
@@ -115,9 +115,9 @@ import (
 	read7 "github.com/devtron-labs/devtron/pkg/module/read"
 	"github.com/devtron-labs/devtron/pkg/module/repo"
 	"github.com/devtron-labs/devtron/pkg/module/store"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication"
-	read9 "github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read"
-	config3 "github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read/config"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication"
+	read9 "github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read"
+	config3 "github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read/config"
 	"github.com/devtron-labs/devtron/pkg/pipeline"
 	"github.com/devtron-labs/devtron/pkg/policyGovernance/security/scanTool"
 	repository11 "github.com/devtron-labs/devtron/pkg/policyGovernance/security/scanTool/repository"
