+++ pkg/appStore/installedApp/service/FullMode/resource/ResourceTreeService.go
@@ -37,12 +37,12 @@ import (
 	"github.com/devtron-labs/devtron/pkg/appStore/bean"
 	appStoreDiscoverRepository "github.com/devtron-labs/devtron/pkg/appStore/discover/repository"
 	"github.com/devtron-labs/devtron/pkg/appStore/installedApp/repository"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication"
+	bean3 "github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/bean"
 	"github.com/devtron-labs/devtron/pkg/deployment/common"
 	bean2 "github.com/devtron-labs/devtron/pkg/deployment/common/bean"
 	"github.com/devtron-labs/devtron/pkg/k8s"
 	application3 "github.com/devtron-labs/devtron/pkg/k8s/application"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication"
-	bean3 "github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/bean"
 	util3 "github.com/devtron-labs/devtron/pkg/util"
 	util2 "github.com/devtron-labs/devtron/util"
 	"github.com/tidwall/gjson"
