+++ pkg/coreEntities/argoApplication/ArgoApplicationService.go
@@ -27,13 +27,13 @@ import (
 	util2 "github.com/devtron-labs/devtron/internal/util"
 	"github.com/devtron-labs/devtron/pkg/cluster/adapter"
 	clusterRepository "github.com/devtron-labs/devtron/pkg/cluster/repository"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/bean"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/helper"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read/config"
 	"github.com/devtron-labs/devtron/pkg/deployment/common"
 	commonBean "github.com/devtron-labs/devtron/pkg/deployment/common/bean"
 	"github.com/devtron-labs/devtron/pkg/k8s/application"
 	k8s2 "github.com/devtron-labs/devtron/pkg/k8s/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/helper"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read/config"
 	"github.com/devtron-labs/devtron/util"
 	"github.com/devtron-labs/devtron/util/sliceUtil"
 	"go.uber.org/zap"
