+++ pkg/coreEntities/argoApplication/ArgoApplicationServiceExtended.go
@@ -29,8 +29,8 @@ import (
 	util2 "github.com/devtron-labs/devtron/internal/util"
 	"github.com/devtron-labs/devtron/pkg/app/appDetails/adapter"
 	"github.com/devtron-labs/devtron/pkg/cluster"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/bean"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read"
 	util3 "github.com/devtron-labs/devtron/pkg/util"
 	"github.com/devtron-labs/devtron/util"
 	"google.golang.org/grpc"
