+++ pkg/coreEntities/argoApplication/bean/bean.go
