+++ pkg/coreEntities/argoApplication/helper/deploymentStatusHelper.go
