+++ pkg/coreEntities/argoApplication/helper/helper.go
@@ -5,7 +5,7 @@ import (
 	"github.com/devtron-labs/common-lib/utils/k8s/commonBean"
 	"github.com/devtron-labs/devtron/api/helm-app/gRPC"
 	"github.com/devtron-labs/devtron/pkg/cluster/repository"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/bean"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/bean"
 	"strconv"
 	"strings"
 )
