+++ pkg/coreEntities/argoApplication/read/ArgoApplicationReadService.go
@@ -26,9 +26,9 @@ import (
 	"github.com/devtron-labs/devtron/api/helm-app/service"
 	"github.com/devtron-labs/devtron/pkg/cluster/adapter"
 	clusterRepository "github.com/devtron-labs/devtron/pkg/cluster/repository"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/bean"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/helper"
 	clientErrors "github.com/devtron-labs/devtron/pkg/errors"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/helper"
 	"go.uber.org/zap"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
