+++ pkg/coreEntities/argoApplication/read/config/ArgoApplicationConfigService.go
@@ -8,8 +8,8 @@ import (
 	k8sCommonBean "github.com/devtron-labs/common-lib/utils/k8s/commonBean"
 	"github.com/devtron-labs/devtron/pkg/cluster/adapter"
 	clusterRepository "github.com/devtron-labs/devtron/pkg/cluster/repository"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/helper"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/bean"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/helper"
 	"go.uber.org/zap"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
