+++ pkg/deployment/deployedApp/status/resourceTree/ResourceTreeService.go
@@ -35,11 +35,11 @@ import (
 	"github.com/devtron-labs/devtron/pkg/app"
 	"github.com/devtron-labs/devtron/pkg/appStatus"
 	read2 "github.com/devtron-labs/devtron/pkg/cluster/environment/read"
+	argoApplication2 "github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication"
+	bean2 "github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/bean"
 	commonBean "github.com/devtron-labs/devtron/pkg/deployment/common/bean"
 	"github.com/devtron-labs/devtron/pkg/k8s"
 	application2 "github.com/devtron-labs/devtron/pkg/k8s/application"
-	argoApplication2 "github.com/devtron-labs/devtron/pkg/nucleus/argoApplication"
-	bean2 "github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/bean"
 	util2 "github.com/devtron-labs/devtron/util"
 	"go.uber.org/zap"
 	"strconv"
