+++ pkg/k8s/K8sCommonService.go
@@ -28,9 +28,9 @@ import (
 	bean2 "github.com/devtron-labs/devtron/pkg/cluster/bean"
 	bean4 "github.com/devtron-labs/devtron/pkg/cluster/environment/bean"
 	"github.com/devtron-labs/devtron/pkg/cluster/read"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read/config"
 	bean3 "github.com/devtron-labs/devtron/pkg/k8s/application/bean"
 	bean5 "github.com/devtron-labs/devtron/pkg/k8s/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read/config"
 	"github.com/devtron-labs/devtron/util"
 	"go.opentelemetry.io/otel"
 	"go.uber.org/zap"
