+++ pkg/k8s/application/k8sApplicationService.go
@@ -28,11 +28,11 @@ import (
 	"github.com/devtron-labs/devtron/pkg/auth/authorisation/casbin"
 	bean5 "github.com/devtron-labs/devtron/pkg/cluster/environment/bean"
 	"github.com/devtron-labs/devtron/pkg/cluster/read"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/helper"
 	clientErrors "github.com/devtron-labs/devtron/pkg/errors"
 	"github.com/devtron-labs/devtron/pkg/fluxApplication"
 	bean2 "github.com/devtron-labs/devtron/pkg/fluxApplication/bean"
 	bean4 "github.com/devtron-labs/devtron/pkg/k8s/bean"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/helper"
 	"io"
 	v1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"net/http"
