+++ wire_gen.go
@@ -227,9 +227,9 @@ import (
 	"github.com/devtron-labs/devtron/pkg/module/repo"
 	"github.com/devtron-labs/devtron/pkg/module/store"
 	"github.com/devtron-labs/devtron/pkg/notifier"
-	"github.com/devtron-labs/devtron/pkg/nucleus/argoApplication"
-	read22 "github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read"
-	config2 "github.com/devtron-labs/devtron/pkg/nucleus/argoApplication/read/config"
+	"github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication"
+	read22 "github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read"
+	config2 "github.com/devtron-labs/devtron/pkg/coreEntities/argoApplication/read/config"
 	"github.com/devtron-labs/devtron/pkg/pipeline"
 	"github.com/devtron-labs/devtron/pkg/pipeline/executors"
 	"github.com/devtron-labs/devtron/pkg/pipeline/history"
